Consolidate writable directories
